Moved to maven-based compilation.
