+++ src/main/java/org/ttrzcinski/dictionaries/Binding.java
