+++ src/main/java/org/ttrzcinski/utils/GitUtil.java
