+++ src/main/java/org/ttrzcinski/utils/OSInfo.java
@@ -1,7 +1,6 @@
 package org.ttrzcinski.utils;
 
-import java.awt.Dimension;
-import java.awt.Toolkit;
+import java.awt.*;
 import java.nio.file.Paths;
 
 /**
