+++ src/main/java/org/ttrzcinski/utils/UUIDUtil.java
